+++ config/environments/development.rb
@@ -28,7 +28,7 @@ Rails.application.configure do
   config.action_mailer.delivery_method = :letter_opener_web
   config.action_mailer.perform_deliveries = true
   config.action_mailer.default_url_options = { host: "localhost", port: 3000 }
-  config.action_mailer.default_options = { from: "support@hamzissystems.com" }
+  config.action_mailer.default_options = { from: "support@example.com" }
 
   config.active_job.queue_adapter = :solid_queue